% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_student.R
\name{data_plot_all}
\alias{data_plot_all}
\title{Plot all pairs of columns in a dataframe}
\usage{
data_plot_all()
}
\arguments{
\item{df}{The dataframe to use for plotting.}
}
\value{
A list of ggplot objects.
}
\description{
This function takes each pair of columns in a dataframe and generates 
scatterplots, boxplots, and histograms. The plots are colorblind-friendly.
}
\examples{
data_plot_all(mtcars)
}
