% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_student.R
\name{data_plot_res}
\alias{data_plot_res}
\title{Plot the residuals of a model}
\usage{
data_plot_res()
}
\arguments{
\item{df}{The dataframe containing residuals.}

\item{residual_col}{The name of the column containing residuals (default is 'residuals').}
}
\value{
A list of ggplot objects (histogram and scatter plot).
}
\description{
This function takes a dataframe with residuals and plots them using ggplot2.
It creates a histogram and scatter plot to evaluate the distribution of the residuals.
}
\examples{
data_plot_res(df1)
}
