% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_student.R
\name{design}
\alias{design}
\title{Fit a statistical model interactively}
\usage{
design()
}
\description{
This function prompts the user to input a dataframe name, selects variables,
fits a mixed-effects model (or other types of models), and optionally saves
the residuals and fitted values to the dataframe and new variables.
}
\examples{
design()
}
